### 2022-12-03 #562 @Polkas
cat2cat: Handling an Inconsistently Coded Categorical Variable in a Panel Dataset
Submitting Author Name: Maciej Nasinski
Submitting Author Github Handle: @polkas
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/Polkas/cat2cat
Submission type: Pre-submission

---

```
Package: cat2cat
Title: Handling an Inconsistently Coded Categorical Variable in a Panel Dataset
Version: 0.4.5.9000
Authors@R: person("Maciej", "Nasinski", email = "nasinski.maciej@gmail.com", role = c("aut", "cre"))
Maintainer: Maciej Nasinski <nasinski.maciej@gmail.com>
  Unifying of an inconsistently coded categorical variable between two different time points in accordance with a mapping table.
  The main rule is to replicate the observation if it could be assign to a few categories.
  Then using simple frequencies or modern statistical methods to approximate probabilities of being assign to each of them.
  This novel procedure was invented and implemented in the paper by (Nasinski, Majchrowska and Broniatowska (2020) <doi:10.24425/cejeme.2020.134747>).
    MASS
    caret,
    randomForest,
    pacman,
    dplyr
```

The main objective is to unify the inconsistently coded categorical variables in a panel/longitudinal dataset.
The supervised methods can be used in the cat2cat procedure.
The output from the cat2cat function can be used in the e.g. weighted linear regression or to assess the counts over the time.

I plan to apply it when know if I can submit the package.

Any scientific field where the panel/longitudinal dataset can be used.
Examples of a panel dataset with such inconsistent coded categorical variables are ones linked with the The International Standard Classification of Occupations (ISCO) and the International Classification of Diseases (ICS).

According to best of my knowledge there is no alternative to my solution other than aggregate the datasets (with some simplifications) or remove the variable.
