+++ tests/testthat/test-plot.R
@@ -1,4 +1,5 @@
 data("occup_small", package = "cat2cat")
+
 occup_old <- occup_small[occup_small$year == 2008, ]
 occup_new <- occup_small[occup_small$year == 2010, ]
 
@@ -9,10 +10,14 @@ occup_2 <- cat2cat(
   mappings = list(trans = trans, direction = "backward")
 )
 
-expect_error(plot_c2c(occup_2$new, type = c("both")))
-expect_error(plot_c2c(occup_2$old, type = c("wrong")))
-expect_error(plot_c2c(NULL, type = c("both")))
+testthat::test_that("plot_c2c wrong input", {
+  expect_error(plot_c2c(occup_2$new, type = c("both")))
+  expect_error(plot_c2c(occup_2$old, type = c("wrong")))
+  expect_error(plot_c2c(NULL, type = c("both")))
+})
 
-expect_silent(plot_c2c(occup_2$old, type = c("both")))
-expect_silent(plot_c2c(occup_2$old, type = c("hist")))
-expect_silent(plot_c2c(occup_2$old, type = c("bar")))
+testthat::test_that("plot_c2c correct input", {
+  expect_silent(plot_c2c(occup_2$old, type = c("both")))
+  expect_silent(plot_c2c(occup_2$old, type = c("hist")))
+  expect_silent(plot_c2c(occup_2$old, type = c("bar")))
+})
