+++ tests/testthat/setup.R
@@ -1,0 +1,1 @@
+library("dplyr")
