+++ tests/testthat/test-lms.R
@@ -30,28 +30,30 @@ ml <- list(
 )
 
 testthat::test_that(
-  "summary_c2c is properly adjust the std error - cat2cat case", {
-  occup <- cat2cat(
-    data = data_simple,
-    mappings = mappings_simple_back,
-    ml = ml
-  )
+  "summary_c2c is properly adjust the std error - cat2cat case",
+  {
+    occup <- cat2cat(
+      data = data_simple,
+      mappings = mappings_simple_back,
+      ml = ml
+    )
 
-  lms <- lm(
-    I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup$old,
-    weights = multiplier * wei_freq_c2c
-  )
+    lms <- lm(
+      I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup$old,
+      weights = multiplier * wei_freq_c2c
+    )
 
-  ss_c2c <- summary_c2c(lms, df_old = nrow(occup_old) - length(lms$assign))
-  lms$df.residual <- nrow(occup_old) - length(lms$assign)
-  ss1 <- suppressWarnings(summary(lms))
+    ss_c2c <- summary_c2c(lms, df_old = nrow(occup_old) - length(lms$assign))
+    lms$df.residual <- nrow(occup_old) - length(lms$assign)
+    ss1 <- suppressWarnings(summary(lms))
 
-  lms2 <- lm(
-    I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup_old,
-    weights = multiplier
-  )
-  ss2 <- summary(lms2)
+    lms2 <- lm(
+      I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup_old,
+      weights = multiplier
+    )
+    ss2 <- summary(lms2)
 
-  expect_true(sum((ss2$coefficients[, 2] - ss1$coefficients[, 2])**2) < 0.01)
-  expect_true(sum((ss2$coefficients[, 2] - ss_c2c$std.error_c)**2) < 0.01)
-})
+    expect_true(sum((ss2$coefficients[, 2] - ss1$coefficients[, 2])**2) < 0.01)
+    expect_true(sum((ss2$coefficients[, 2] - ss_c2c$std.error_c)**2) < 0.01)
+  }
+)
