+++ tests/testthat/test-cat2cat-utils.R
@@ -24,51 +24,54 @@ testthat::test_that("dummy_c2c backward", {
   )
 })
 
-###
-
-# cat2cat utils
-library("dplyr")
-set.seed(1234)
-
 data("occup", package = "cat2cat")
 data("trans", package = "cat2cat")
 
 occup_old <- occup[occup$year == 2008, ]
 occup_new <- occup[occup$year == 2010, ]
 
-c2c <- cat2cat(
-  data = list(
-    old = occup_old,
-    new = occup_new,
-    cat_var = "code",
-    time_var = "year",
-    multiplier_var = "multiplier"
-  ),
-  mappings = list(trans = trans, direction = "backward")
-)
+testthat::test_that("prune_c2c on cat2cat output", {
+  c2c <- cat2cat(
+    data = list(
+      old = occup_old,
+      new = occup_new,
+      cat_var = "code",
+      time_var = "year",
+      multiplier_var = "multiplier"
+    ),
+    mappings = list(trans = trans, direction = "backward")
+  )
 
-expect_equal(sum(c2c$old$wei_freq_c2c), nrow(occup_old))
-expect_true(all(c2c$old$wei_freq_c2c <= 1 & c2c$old$wei_freq_c2c >= 0))
-expect_identical(
-  nrow(occup_old),
-  c2c$old %>%
-    prune_c2c(method = "highest1") %>%
-    nrow()
-)
-
-expect_true(
-  (c2c$old %>%
-    prune_c2c(method = "highest") %>%
-    nrow()) >=
-    (c2c$old %>%
+  expect_equal(sum(c2c$old$wei_freq_c2c), nrow(occup_old))
+  expect_true(all(c2c$old$wei_freq_c2c <= 1 & c2c$old$wei_freq_c2c >= 0))
+  expect_identical(
+    nrow(occup_old),
+    c2c$old %>%
       prune_c2c(method = "highest1") %>%
-      nrow())
-)
+      nrow()
+  )
 
-expect_equal(
-  nrow(occup_old),
-  c2c$old %>%
-    prune_c2c(method = "morethan", percent = 0.2) %>%
-    pull("wei_freq_c2c") %>%
-    sum()
-)
+  expect_identical(
+    sum(c2c$old$wei_freq_c2c > 0),
+    c2c$old %>%
+      prune_c2c(method = "nonzero") %>%
+      nrow()
+  )
+
+  expect_true(
+    (c2c$old %>%
+      prune_c2c(method = "highest") %>%
+      nrow()) >=
+      (c2c$old %>%
+        prune_c2c(method = "highest1") %>%
+        nrow())
+  )
+
+  expect_equal(
+    nrow(occup_old),
+    c2c$old %>%
+      prune_c2c(method = "morethan", percent = 0.2) %>%
+      pull("wei_freq_c2c") %>%
+      sum()
+  )
+})
