+++ tests/testthat/test-cat2cat.R
@@ -1,4 +1,3 @@
-suppressPackageStartupMessages(library("dplyr"))
 set.seed(1234)
 
 data("occup", package = "cat2cat")

[…]
