+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.6.9000
 pkgdown_sha: 977a429ed34c4b76e0982dfc2d656bcae159d81f
 articles:
   cat2cat: cat2cat.html
-last_built: 2022-12-03T19:16Z
+last_built: 2022-12-04T18:29Z
 
