### 2022-12-06 #562 @annakrystalli
Dear @Polkas,

The editorial team has concluded that the package definitely fits in our "stats" scope.

Before proceeding and closing this pre-sub enquiry, there is also a need to clarify what category it would fit. The [stats-devguide states categories are appropriate where at least half of all standards can be applied](https://stats-devguide.ropensci.org/pkgdev.html#scope). We suggest you need to try and narrow down to one category only.

We feel it does not best fit the "time series" category and seems initially to most likely be "Machine Learning," We suggest  you spend a little time to read though the standards and consider which you would think most appropriate.

Following that, the best way to confirm would be to go through the formal process of documenting compliance with the stats standards, which needs to be done prior to submission anyways.  You can call `@ropensci-review-bot check srr` in this issue to confirm documentation has been completed successfully. You can find [more details in our documentation](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr).

Just ping me here to confirm that's done and the category you have narrowed it down too or if you need any help.

Thanks again for your enquiry!
