+++ DESCRIPTION
@@ -4,10 +4,10 @@ Version: 0.4.5.9000
 Authors@R: person("Maciej", "Nasinski", email = "nasinski.maciej@gmail.com", role = c("aut", "cre"))
 Maintainer: Maciej Nasinski <nasinski.maciej@gmail.com>
 Description: 
-  Unifying of an inconsistently coded categorical variable between two different time points in accordance with a mapping table.
-  The main rule is to replicate the observation if it could be assign to a few categories.
-  Then using simple frequencies or modern statistical methods to approximate probabilities of being assign to each of them.
-  This novel procedure was invented and implemented in the paper by (Nasinski, Majchrowska and Broniatowska (2020) <doi:10.24425/cejeme.2020.134747>).
+    Unifying an inconsistently coded categorical variable between two different time points in accordance with a mapping table. 
+    The main rule is to replicate the observation if it could be assigned to a few categories. 
+    Then using frequencies or statistical methods to approximate the probabilities of being assigned to each of them. 
+    This procedure was invented and implemented in the paper by (Nasinski, Majchrowska, and Broniatowska (2020) <doi:10.24425/cejeme.2020.134747>).
 Depends: R (>= 3.6)
 License: GPL (>= 2)
 URL: https://github.com/Polkas/cat2cat, https://polkas.github.io/cat2cat/
