+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.19.2
-pkgdown: 2.0.6.9000
-pkgdown_sha: 977a429ed34c4b76e0982dfc2d656bcae159d81f
+pkgdown: 2.0.7.9000
+pkgdown_sha: c354aa7e5ea1f9936692494c28c89e5bdd31fc68
 articles:
   cat2cat: cat2cat.html
-last_built: 2022-12-04T18:29Z
+last_built: 2023-01-25T14:23Z
 
