+++ README.md
@@ -1,7 +1,7 @@
 # cat2cat <a href='https://github.com/polkas/cat2cat'><img src='man/figures/cat2cat_logo.png' align="right" width="200px" /></a>
 [![R build status](https://github.com/polkas/cat2cat/workflows/R-CMD-check/badge.svg)](https://github.com/polkas/cat2cat/actions)
 [![CRAN](http://www.r-pkg.org/badges/version/cat2cat)](https://cran.r-project.org/package=cat2cat)
-[![codecov](https://codecov.io/gh/Polkas/cat2cat/branch/main/graph/badge.svg)](https://codecov.io/gh/Polkas/cat2cat)
+[![codecov](https://codecov.io/gh/Polkas/cat2cat/branch/main/graph/badge.svg)](https://app.codecov.io/gh/Polkas/cat2cat)
 [![Dependencies](https://tinyverse.netlify.com/badge/cat2cat)](https://cran.r-project.org/package=cat2cat)
 
 ## Handling an Inconsistent Coded Categorical Variable in a Panel Dataset
@@ -10,7 +10,7 @@ Unifying an inconsistent coded categorical variable in a panel/longtitudal datas
 There is offered the novel `cat2cat` procedure to map a categorical variable according to a mapping (transition) table between two different time points.
 The mapping (transition) table should to have a candidate for each category from the targeted for an update period. The main rule is to replicate the observation if it could be assigned to a few categories, then using simple frequencies or modern statistical methods to approximate probabilities of being assigned to each of them.
 
-**This algorithm was invented and implemented in the paper by (Nasinski, Majchrowska and Broniatowska (2020) <doi:10.24425/cejeme.2020.134747>).**
+**This algorithm was invented and implemented in the paper by (Nasinski, Majchrowska and Broniatowska (2020) <\doi:10.24425/cejeme.2020.134747>).**
 
 [**Please visit the cat2cat webpage for more information**](https://polkas.github.io/cat2cat/articles/cat2cat.html)
 
