+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.19.2
-pkgdown: 2.0.7.9000
-pkgdown_sha: c354aa7e5ea1f9936692494c28c89e5bdd31fc68
+pkgdown: 2.0.7
+pkgdown_sha: ~
 articles:
   cat2cat: cat2cat.html
-last_built: 2023-01-30T18:55Z
+last_built: 2023-01-30T22:36Z
 
