### 2023-02-01 #562 @maurolepore
Dear @Polkas,

Today starts my rotation as EiC meaning the role of @annakrystalli is now mine. Did you have the chance to follow up on the comment above?
