+++ DESCRIPTION
@@ -26,5 +26,5 @@ Suggests:
     dplyr
 LazyData: true
 VignetteBuilder: knitr
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.2
 Config/testthat/edition: 3
