+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   cat2cat: cat2cat.html
-last_built: 2023-01-30T22:36Z
+last_built: 2023-02-01T19:31Z
 
