### 2023-05-23 #562 @maelle
@Polkas friendly reminder, did you get a chance to work on the comments from https://github.com/ropensci/software-review/issues/562#issuecomment-1339049291?
