### 2023-12-15 #562 @jhollist
@Polkas I am currently serving as the EIC and am checking in on some older submissions.  First, congrats on the publication!  You mentioned that you might pursue another submission to rOpenSci.  Have you decided to move forward with that?
