### 2023-12-17 #562 @Polkas
Hey @jhollist, thank you for your response. I have dedicated effort to align with the expected standards. However, it appears that the current focus of rOpenSci may have shifted away from packages similar to mine.

I understand that rOpenSci is now prioritizing support for packages that facilitate reproducible research and manage the data lifecycle for scientists. I have thoroughly reviewed the current [package categories](https://devguide.ropensci.org/policies.html#package-categories) and, unfortunately, it seems my package may not align with any of these categories.

If my understanding is correct and my package indeed falls outside the scope of rOpenSci's current focus, please feel free to close this issue.
