### 2023-12-18 #562 @jhollist
@Polkas your package is a better fit for our Statistical Software.  Based on the conversation above (https://github.com/ropensci/software-review/issues/562#issuecomment-1339049291), take a close look at https://stats-devguide.ropensci.org/pkgdev.html#scope and see if you think any of those fit.  The prior conversations on here and amongst the editors felt like Machine Learning might be the best fit.  If you would like to proceed take a close look at the Stats devguide.  If you have specific questions after that, you can ping me again here.  Thanks!
