### 2024-02-28 #562 @ldecicco-USGS
Hi @Polkas ! I'm checking in on submissions that have been sitting for awhile. It sounds like the feedback has been that this package would be better suited for the rOpenSci Statistical Software submission. The process is similar, but there are a few differences. I'll once again plug the statistical submission guide:

https://stats-devguide.ropensci.org/

Let me know if you have any questions.
