### 2024-08-24 #562 @adamhsparks
Hi @Polkas, just following up on your submission. Are you interested in submitting this package for review as a statistical package? We'd be happy to review it. Please see [stats-devguide.ropensci.org](https://stats-devguide.ropensci.org/) for more.

If you don't wish to proceed, could you let us know so we can close this issue?
