### 2024-10-15 #562 @adamhsparks
Hi @Polkas, I'm just checking in to see if you were interested in submitting this package as a statistical package for review? If not, I'll go ahead and close this issue by the end of October.
